Refactor recheckable Future
