+++ crates/medea-reactive/src/collections/hash_map.rs
@@ -8,10 +8,7 @@ use std::{
     marker::PhantomData,
 };
 
-use futures::{
-    future::{self, FutureExt as _, LocalBoxFuture},
-    stream::{LocalBoxStream, StreamExt as _},
-};
+use futures::stream::{LocalBoxStream, StreamExt as _};
 
 use crate::subscribers_store::{
     common, progressable,
