+++ crates/medea-reactive/src/collections/hash_set.rs
@@ -2,9 +2,7 @@
 
 use std::{collections::hash_set::Iter, hash::Hash, marker::PhantomData};
 
-use futures::{
-    future, future::LocalBoxFuture, stream::LocalBoxStream, FutureExt,
-};
+use futures::stream::LocalBoxStream;
 
 use crate::subscribers_store::{
     common, progressable,
